Handle GONE status from Watch() by re-syncing
On occasion, the k8s API returns `GONE` (410) when attempting to
re-establish a watch after disconnecting from the control-plane.

In that case, we need to get a full dump of all existing pods and
generate events for new, existing and dead pods.

Example error we'll be attempting to recover from:

```
&Status{ListMeta:ListMeta{SelfLink:,ResourceVersion:,Continue:,RemainingItemCount:nil,},Status:Failure,Message:too old resource version: 165523336 (173189307),Reason:Gone,Details:nil,Code:410,}
```

Add a trivial test for the new tracker code.