PodWatcher: Add Continues method to PodEvent
Most use-cases require knowledge of when the initial state-dump (from a
List call) has completed before doing some initialization.

PodWatcher was doing almost all the accounting necessary for this to
work but not exposing any definition between the last CreatePod event
of the initial state-dump and anything delivered by the watcher. (which
may happen arbitrarily far in the future).

Solve this problem by adding a Continues() method to PodEvent (and its
implementations) and adjusting watch(), resync() and initialPods() to
set the new field appropriately.

Note: this change adds an unexported setContinues function to set the
continues field on all three event types because I want to keep the
PodEvent interface implementable by other packages to allow for easier
testing of the callbacks. (which would be precluded by an unexported
method for setting that field)