pod watcher: Add InitialListComplete event
In some cases, the initial list of pods matching the specified label
matcher may be empty. In that case, clients may still want to continue
initialization. However, the `Continues()` method only provides
information about the first full state if there are pod events to attach
them to.

This change adds a `InitialListComplete` event type that lets a caller
figure out when the initial state dump (of `CreatePod` events) is
complete even when there are no pods that matched the label-matcher.